requirements.txt
addict==2.4.0
alembic==1.13.1
appnope==0.1.4
asteval==0.9.31
asttokens==2.4.1
attrs==23.2.0
banal==1.0.6
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
ConfigArgParse==1.7
contourpy==1.2.0
cycler==0.12.1
dash==2.15.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataset==1.6.2
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
Flask==3.0.2
fonttools==4.49.0
fsspec==2024.2.0
future==1.0.0
graphviz==0.20.1
greenlet==3.0.3
gtbook==0.0.37
gtsam==4.2
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
install==1.3.5
ipykernel==6.29.2
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kaleido==0.2.1
kiwisolver==1.4.5
lmfit==1.2.2
Mako==1.3.2
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mpmath==1.3.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
open3d==0.18.0
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.0
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
plotly==5.19.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.1
pyquaternion==0.9.9
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
requests==2.31.0
retrying==1.3.4
rpds-py==0.18.0
scikit-learn==1.4.1.post1
scipy==1.12.0
simpleicp==2.0.14
six==1.16.0
SQLAlchemy==1.4.51
stack-data==0.6.3
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.3.0
torch==2.2.0
torchvision==0.17.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transforms3d==0.4.1
typing_extensions==4.9.0
tzdata==2024.1
uncertainties==3.1.7
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.1
zipp==3.17.0
